应用协调层/middle/requirements.txt
# ============================================================
# 应用协调层 - LangChain 混合检索系统
# ============================================================
# 功能：混合检索、智能路由、API服务
# 环境：Python 3.11.13
# 更新时间：2025-10-21
# 状态：✅ 已验证稳定运行
# ============================================================

# ============================================================
# 1. LangChain 生态系统
# ============================================================
langchain==0.2.15
langchain-core==0.2.43
langchain-community==0.2.10
langchain-openai==0.1.17
langchain-text-splitters==0.2.2

# ============================================================
# 2. LangGraph
# ============================================================
langgraph==0.2.10
langgraph-checkpoint==1.0.12
langgraph-sdk==0.2.9
langsmith==0.1.147

# ============================================================
# 3. API 框架
# ============================================================
fastapi==0.119.1
uvicorn==0.38.0
starlette==0.48.0
pydantic==2.12.3
pydantic-core==2.41.4
pydantic-settings==2.11.0

# ============================================================
# 4. AI 服务客户端
# ============================================================
openai==1.109.1
anthropic==0.21.3
cohere==4.43
tiktoken==0.7.0

# ============================================================
# 5. 数据库驱动（混合检索需要）
# ============================================================
neo4j==5.25.0
chromadb==1.2.0
chroma-hnswlib==0.7.3
redis==5.0.8

# ============================================================
# 6. Embeddings（与向量数据库一致）
# ============================================================
sentence-transformers==2.3.1
transformers==4.36.2
huggingface-hub==0.24.0
tokenizers==0.15.0

# ============================================================
# 7. HTTP 客户端
# ============================================================
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.13.1
aiofiles==23.2.1
requests==2.32.5

# ============================================================
# 8. 数据处理
# ============================================================
numpy==1.26.4
pandas==2.0.3

# ============================================================
# 9. 异步支持
# ============================================================
tenacity==8.5.0
nest-asyncio==1.6.0
anyio==4.11.0

# ============================================================
# 10. 配置管理
# ============================================================
pyyaml==6.0.3
python-dotenv==1.1.1

# ============================================================
# 11. 日志工具
# ============================================================
loguru==0.7.2
rich==13.7.1
colorama==0.4.6

# ============================================================
# 12. 数据序列化
# ============================================================
orjson==3.11.3
msgpack==1.1.2
jsonpatch==1.33

# ============================================================
# 13. 缓存
# ============================================================
diskcache==5.6.3

# ============================================================
# 14. 中文文本处理（可选）
# ============================================================
jieba==0.42.1

# ============================================================
# 使用说明
# ============================================================
# 
# 安装：
# pip install -r requirements.txt
# 
# 注意：
# - 需要先安装检索与知识层的依赖
# - Neo4j 需要单独启动服务
# - ChromaDB 向量数据库需要已构建
# ============================================================
